packages/gcs-connector/backend/requirements.txt
# Google Cloud Storage Connector Backend Dependencies
# These are merged into the main backend's requirements.txt on install

fastapi>=0.115.0
pydantic>=2.9.0
httpx>=0.27.0
google-cloud-storage>=2.18.0
google-auth>=2.35.0
